reorganized command line flag parsing code
