+++ util.go
@@ -10,6 +10,44 @@ import (
 	"fmt"
 )
 
+// exidCollector accumulates the external ids from the command line -e and -E
+// flags
+var exidCollector [][]byte
+
+// extids will be a flag receiver for adding chains and entries
+// In ASCII
+type extidsAscii []string
+
+func (e *extidsAscii) String() string {
+	return fmt.Sprint(*e)
+}
+
+func (e *extidsAscii) Set(s string) error {
+	*e = append(*e, s)
+	exidCollector = append(exidCollector[:], []byte(s))
+	return nil
+}
+
+// extids will be a flag receiver for adding chains and entries
+// In HEX
+type extidsHex []string
+
+func (e *extidsHex) String() string {
+	return fmt.Sprint(*e)
+}
+
+func (e *extidsHex) Set(s string) error {
+	*e = append(*e, s)
+	b, err := hex.DecodeString(s)
+	if err != nil {
+		return err
+	}
+	exidCollector = append(exidCollector[:], b)
+	return nil
+}
+
+// nameCollector accumulates the components of a chain name from the command
+// line -n and -N flags
 var nameCollector [][]byte
 
 // namesAscii will be a flag receiver for ASCII chain names.
